+++ packages/earth-admin/src/components/dropdown/styles.scss
@@ -66,8 +66,9 @@
 
 .ng-select-list {
   display: block;
+  height: auto;
+  overflow: auto;
   .ng-nav-item {
     display: block;
-
   }
 }
