+++ packages/earth-admin/src/utils/index.ts
@@ -19,6 +19,7 @@
 
 import { navigate } from 'gatsby';
 import { BASE_URL } from 'config';
+import {ADMIN_PAGES} from 'components/sidebar-select/model';
 import get from 'lodash/get';
 import isArray from 'lodash/isArray';
 import isString from 'lodash/isString';
@@ -33,6 +34,7 @@ const DeserializerService = new JSONAPIDeserializer({
 });
 
 import queryStringEncode from 'query-string-encode';
+import {childOfKind} from 'tslint';
 
 /**
  * Wrapper over navigate that takes into account baseURL.
@@ -121,3 +123,8 @@ export const removeNestedGroups = (groups: string[]): string[] => {
 
 export const isValidOrg = (orgsFromToken: string[], org: string): boolean =>
   orgsFromToken.includes(org);
+
+
+export const setPage = (pageType: string) => {
+  return ADMIN_PAGES.filter(page => page.key === pageType);
+}
